extracted implementation part of `Context.h` to `Context.c`
